+++ src/rpc/events/request_event.rs
@@ -7,7 +7,7 @@ pub struct RequestEvent<'a> {
     prevent_default: bool,
     message: &'a dyn MessageBase,
     node: Option<Node>,
-    received_time: u64,
+    received_time: u128,
     response: Option<Box<dyn MessageBase>>
 }
 
@@ -69,11 +69,11 @@ impl<'a> MessageEvent for RequestEvent<'a> {
         self.node.unwrap()
     }
 
-    fn set_received_time(&mut self, received_time: u64) {
+    fn set_received_time(&mut self, received_time: u128) {
         self.received_time = received_time;
     }
 
-    fn get_received_time(&self) -> u64 {
+    fn get_received_time(&self) -> u128 {
         self.received_time
     }
